project initialization, refactoring in multiple apps, migrations prepared
