+++ lettings/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+
+from .models import Letting, Address
+
+admin.site.register(Letting)
+admin.site.register(Address)
