+++ oc_lettings_site/admin.py
@@ -1,10 +1,1 @@
 from django.contrib import admin
-
-from .models import Letting
-from .models import Address
-from .models import Profile
-
-
-admin.site.register(Letting)
-admin.site.register(Address)
-admin.site.register(Profile)
